move getFilterBuilder out of MongoDbPlanner into separate file
